Remove dividers from Translatable .
It is, perhaps, simpler to define dividers in `Display::Panels::X`
instead of passing through from `Translatable`. Which (if any) dividers
to use is part of the design of the panel. One option would be to define
dividers per-`Panel` in `Translatable`; however, if this functionality
is required, the dividers can again be extracted.