Fix Display::Subpanel#content overrun.
When the `Display::Subpanel#content` was too long for the available
space, the content overran. This overrun wasn't necessarily visible,
as other content was written over the top. Where lambda data at the
right of the screen was updated to become too long, however, the overrun
was visible. Fix by forcing a truncation of the content.