[#3] Call #close properly when exception in Display::Screen#initialize .
When an exception occurred in `Display::Screen#initialize`, `#close` did
not get called, meaning Curses-related cleanup did not occur. Fix by
rescuing as part of `#initialize`, calling `#close`, and reraising.

This enables error messages to be output to the screen; otherwise,
nothing shows because of failure to shutdown Curses-related things.