[#5] Implement Spy::Queues, retrieving Sidekiq Stats Queues data.
This provides data equivalent to the `Sidekiq::Web` Queues page (but
not the linked Queue pages). It uses `Sidekiq::Stats#queues`, not
`Sidekiq::Queue`, as only the name and size of the queues are retrieved
(as per the `Sidekiq::Web` Queues page).

This is implemented in a similar way to `Spy::Workers`.