[#5] Define Display::Panels::Queues, equivalent to Queues page.
Adding to the existing skeleton, use the data provided by `Spy::Queues`
to display a list of active Sidekiq queues and their sizes. Note that it
is necessary to hook into `#refresh` and call `Spy::Queues#refresh`. The
list is defined to be precisely as many table rows long as the panel
minus the header row.

Note that the `Sidekiq::Web` Queue page (linked from Queues) is not
implemented as part of this work.