[#6] Implement Spy::Retries, retrieving Sidekiq Retries data.
This provides data equivalent to the `Sidekiq::Web` Retries page (but
not the linked Retry pages). It uses `Sidekiq::RetrySet` with no limit.
Because of this, the current implementation could experience issues if
there are a large number of retries.

There is no sort applied to the list, so it should reflect any
underlying order returned from Sidekiq.

This is implemented in a similar way to `Spy::Workers`.