[#6] Define Display::Panels::Retries, equivalent to Retries page.
Using the data provided by `Spy::Retries`, define panel
`Display::Panels::Retries`. This displays a list of queued Sidekiq
retries. Note that it is necessary to hook into `#refresh` and call
`Spy::Retries#refresh`. The list is defined to be precisely as many
table rows long as the panel minus the header row.

Mount a retries panel in `Display::Screen` in the `:main` panel,
binding to keypress `<r>`. Add a Retries entry to the `:header` panel
menu.

Note that the `Sidekiq::Web` Retry page (linked from Retries) is not
implemented as part of this work.