[#7] Implement Spy::Schedules, retrieving Sidekiq Scheduled data.
This provides data equivalent to the `Sidekiq::Web` Scheduled page. It
uses `Sidekiq::ScheduledSet` with no limit. Because of this, the current
implementation could experience issues if there are a large number of
schedules.

There is no sort applied to the list, so it should reflect any
underlying order returned from Sidekiq.

The name 'Schedules' rather than 'Scheduled' is used for internal
consistency, despite it being named the latter in Sidekiq. I might live
to regret my choice.

This is implemented in a similar way to `Spy::Retries`.