[#7] Define Display::Panels::Schedules, equivalent to Scheduled page.
Using the data provided by `Spy::Schedules`, define panel
`Display::Panels::Schedules`. This displays a list of scheduled jobs.
Note that it is necessary to hook into `#refresh` and call
`Spy::Schedules#refresh`. The list is defined to be precisely as many
table rows long as the panel minus the header row.

Mount a schedules panel in `Display::Screen` in the `:main` panel,
binding to keypress `<s>`. Add a Schedules entry to the `:header` panel
menu, displayed as 'SCHEDULED' (to match `Sidekiq::Web`).