Fix curses for supported Ruby versions, including Ruby 1.9.3 .
d85ea78 installs curses, because curses has been removed from the
standard library in Ruby 2.1.0 onwards. However, the curses gem does
not appear to work with Ruby 1.9.3 . Thus, it cannot be added as a
dependency in `sidekiq-spy.gemspec`.

`Gemfile` supports the notion of platforms. However, `Gemfile` is not
used at build-time. Add curses as a conditional dependency at install-
time, using `ext/mkrf_conf.rb`, based on

    http://en.wikibooks.org/wiki/Ruby_Programming/RubyGems#How_to_install_different_versions_of_gems_depending_on_which_version_of_ruby_the_installee_is_using

However, the dependency still needs to be satisfied at development-
time. Adding this to `Gemfile` using `:platform => :ruby_21` does not
appear to work in Ruby 1.9.3 , because it is an invalid platform there.
Thus, use `RUBY_VERSION` to specify curses as a development dependency
for Ruby 2.1.0 onwards.

If you know a better way of doing this, I would be very interested to
hear it. :)