Extend Config to support Redis username,password .
Add `-U --username` (`-u` is taken), `-P --password` (`-p` is taken)
to `bin/sidekiq-spy`. Extend `Config` to support set both directly
and through `#url=`. Update `#url` to simplify and support settings.

Note that Redis doesn't really have the concept of username, but it's
needed to form a valid URI when password is set

    http://www.w3.org/Addressing/URL/url-spec.txt ,15

so it might as well be exposed rather than populated internally.